# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DealBattle",
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/v1/session": {
      "post": {
        "summary": "Create Session",
        "operationId": "create_session_api_v1_session_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionView"
                }
              }
            }
          },
          "400": {
            "description": "IDEMPOTENCY_REQUIRED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            }
          },
          "401": {
            "description": "SESSION_REQUIRED / CREDENTIAL_INVALID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            }
          },
          "403": {
            "description": "CSRF_INVALID / ORIGIN_INVALID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            }
          },
          "404": {
            "description": "DEAL_NOT_FOUND / PRODUCT_NOT_FOUND",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            }
          },
          "409": {
            "description": "VERSION_CONFLICT / STATE_CONFLICT / APPROVAL_STALE / IDEMPOTENCY_CONFLICT",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            }
          },
          "422": {
            "description": "VALIDATION_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            }
          },
          "503": {
            "description": "DATABASE_UNAVAILABLE",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Get Session",
        "operationId": "get_session_api_v1_session_get",
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionView"
                }
              }
            }
          },
          "400": {
            "description": "IDEMPOTENCY_REQUIRED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            }
          },
          "401": {
            "description": "SESSION_REQUIRED / CREDENTIAL_INVALID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            }
          },
          "403": {
            "description": "CSRF_INVALID / ORIGIN_INVALID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            }
          },
          "404": {
            "description": "DEAL_NOT_FOUND / PRODUCT_NOT_FOUND",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            }
          },
          "409": {
            "description": "VERSION_CONFLICT / STATE_CONFLICT / APPROVAL_STALE / IDEMPOTENCY_CONFLICT",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            }
          },
          "422": {
            "description": "VALIDATION_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            }
          },
          "503": {
            "description": "DATABASE_UNAVAILABLE",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Expire Session",
        "operationId": "expire_session_api_v1_session_delete",
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "parameters": [
          {
            "name": "x-csrf-token",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-Csrf-Token"
            }
          },
          {
            "name": "idempotency-key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Idempotency-Key"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "400": {
            "description": "IDEMPOTENCY_REQUIRED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            }
          },
          "401": {
            "description": "SESSION_REQUIRED / CREDENTIAL_INVALID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            }
          },
          "403": {
            "description": "CSRF_INVALID / ORIGIN_INVALID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            }
          },
          "404": {
            "description": "DEAL_NOT_FOUND / PRODUCT_NOT_FOUND",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            }
          },
          "409": {
            "description": "VERSION_CONFLICT / STATE_CONFLICT / APPROVAL_STALE / IDEMPOTENCY_CONFLICT",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            }
          },
          "422": {
            "description": "VALIDATION_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            }
          },
          "503": {
            "description": "DATABASE_UNAVAILABLE",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/deals": {
      "post": {
        "summary": "Create Deal",
        "operationId": "create_deal_api_v1_deals_post",
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "parameters": [
          {
            "name": "x-csrf-token",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-Csrf-Token"
            }
          },
          {
            "name": "idempotency-key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDeal"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DealView"
                }
              }
            }
          },
          "400": {
            "description": "IDEMPOTENCY_REQUIRED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            }
          },
          "401": {
            "description": "SESSION_REQUIRED / CREDENTIAL_INVALID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            }
          },
          "403": {
            "description": "CSRF_INVALID / ORIGIN_INVALID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            }
          },
          "404": {
            "description": "DEAL_NOT_FOUND / PRODUCT_NOT_FOUND",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            }
          },
          "409": {
            "description": "VERSION_CONFLICT / STATE_CONFLICT / APPROVAL_STALE / IDEMPOTENCY_CONFLICT",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            }
          },
          "422": {
            "description": "VALIDATION_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            }
          },
          "503": {
            "description": "DATABASE_UNAVAILABLE",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AgreementView": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "snapshot": {
            "$ref": "#/components/schemas/Snapshot"
          },
          "snapshot_hash": {
            "type": "string",
            "title": "Snapshot Hash"
          },
          "valid": {
            "type": "boolean",
            "title": "Valid"
          },
          "approval": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approval"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "snapshot",
          "snapshot_hash",
          "valid",
          "approval"
        ],
        "title": "AgreementView"
      },
...
      "BuyerPolicy": {
        "properties": {
          "max_total_krw": {
            "type": "integer",
            "maximum": 100000000.0,
            "exclusiveMinimum": 0.0,
            "title": "Max Total Krw"
          },
          "min_ram_gb": {
            "type": "integer",
            "maximum": 256.0,
            "minimum": 1.0,
            "title": "Min Ram Gb"
          },
          "min_ssd_gb": {
            "type": "integer",
            "maximum": 8192.0,
            "minimum": 1.0,
            "title": "Min Ssd Gb"
          },
          "allowed_seller_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 20,
            "minItems": 1,
            "title": "Allowed Seller Ids"
          },
          "delivery_by": {
            "type": "string",
            "format": "date",
            "title": "Delivery By"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expires At"
          },
          "max_rounds": {
            "type": "integer",
            "maximum": 12.0,
            "minimum": 2.0,
            "title": "Max Rounds",
            "default": 6
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "max_total_krw",
          "min_ram_gb",
          "min_ssd_gb",
          "allowed_seller_ids",
          "delivery_by",
          "expires_at"
        ],
        "title": "BuyerPolicy"
      },
      "ChainView": {
        "properties": {
          "record_id": {
            "type": "string",
            "title": "Record Id"
          },
          "adapter_mode": {
            "type": "string",
            "enum": [
              "mock",
              "evm"
            ],
            "title": "Adapter Mode"
          },
          "status": {
            "type": "string",
            "enum": [
              "QUEUED",
              "SUBMITTING",
              "PENDING",
              "CONFIRMED",
              "MOCK_RECORDED",
              "UNKNOWN",
              "FAILED"
            ],
            "title": "Status"
          },
          "audit_hash": {
            "type": "string",
            "title": "Audit Hash"
          },
          "chain_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chain Id"
          },
          "contract": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contract"
          },
          "tx_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tx Hash"
          },
          "nonce": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nonce"
          },
          "explorer_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Explorer Url"
          },
          "receipt": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Receipt"
          },
          "event": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Event"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "record_id",
          "adapter_mode",
          "status",
          "audit_hash",
          "chain_id",
          "contract",
          "tx_hash",
          "nonce",
          "explorer_url",
          "receipt",
          "event"
        ],
        "title": "ChainView"
      },
...
      "CreateDeal": {
        "properties": {
          "product_id": {
            "type": "string",
            "title": "Product Id"
          },
          "policy": {
            "$ref": "#/components/schemas/BuyerPolicy"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "product_id",
          "policy"
        ],
        "title": "CreateDeal"
      },
      "DealView": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "mode": {
            "type": "string",
            "enum": [
              "mock",
              "live"
            ],
            "title": "Mode"
          },
          "status": {
            "$ref": "#/components/schemas/State"
          },
          "product": {
            "$ref": "#/components/schemas/ProductView"
          },
          "policy": {
            "$ref": "#/components/schemas/BuyerPolicy"
          },
          "policy_version": {
            "type": "integer",
            "title": "Policy Version"
          },
          "policy_confirmed": {
            "type": "boolean",
            "title": "Policy Confirmed"
          },
          "reason_codes": {
            "items": {
              "$ref": "#/components/schemas/Reason"
            },
            "type": "array",
            "title": "Reason Codes"
          },
          "agreement": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AgreementView"
              },
              {
                "type": "null"
              }
            ]
          },
          "chain": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ChainView"
              },
              {
                "type": "null"
              }
            ]
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "mode",
          "status",
          "product",
          "policy",
          "policy_version",
          "policy_confirmed",
          "reason_codes",
          "agreement",
          "chain",
          "created_at"
        ],
        "title": "DealView"
      },
...
      "ErrorDetail": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorDetail"
      },
      "ErrorView": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorView"
      },
...
      "ProductView": {
        "properties": {
          "product_id": {
            "type": "string",
            "title": "Product Id"
          },
          "seller_id": {
            "type": "string",
            "title": "Seller Id"
          },
          "product_version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Product Version"
          },
          "seller_version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Seller Version"
          },
          "name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Name"
          },
          "ram_gb": {
            "type": "integer",
            "maximum": 256.0,
            "minimum": 1.0,
            "title": "Ram Gb"
          },
          "ssd_gb": {
            "type": "integer",
            "maximum": 8192.0,
            "minimum": 1.0,
            "title": "Ssd Gb"
          },
          "asking_price_krw": {
            "type": "integer",
            "maximum": 100000000.0,
            "exclusiveMinimum": 0.0,
            "title": "Asking Price Krw"
          },
          "shipping_fee_krw": {
            "type": "integer",
            "maximum": 100000000.0,
            "minimum": 0.0,
            "title": "Shipping Fee Krw"
          },
          "fee_krw": {
            "type": "integer",
            "maximum": 100000000.0,
            "minimum": 0.0,
            "title": "Fee Krw"
          },
          "delivery_date": {
            "type": "string",
            "format": "date",
            "title": "Delivery Date"
          },
          "stock": {
            "type": "integer",
            "title": "Stock"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expires At"
          },
          "source": {
            "type": "string",
            "const": "simulated",
            "title": "Source"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "product_id",
          "seller_id",
          "product_version",
          "seller_version",
          "name",
          "ram_gb",
          "ssd_gb",
          "asking_price_krw",
          "shipping_fee_krw",
          "fee_krw",
          "delivery_date",
          "stock",
          "expires_at",
          "source"
        ],
        "title": "ProductView"
      },
...
      "Reason": {
        "type": "string",
        "enum": [
          "BUDGET_EXCEEDED",
          "SELLER_FLOOR_VIOLATED",
          "SELLER_NOT_ALLOWED",
          "SPEC_MISMATCH",
          "DELIVERY_TOO_LATE",
          "OUT_OF_STOCK",
          "POLICY_EXPIRED",
          "PRODUCT_EXPIRED",
          "ROUND_LIMIT",
          "INVALID_MODEL_OUTPUT",
          "MODEL_UNAVAILABLE",
          "USAGE_UNAVAILABLE",
          "MODEL_REJECTED",
          "ACCEPT_MISMATCH",
          "APPROVAL_STALE",
          "INVALID_DATA"
        ],
        "title": "Reason"
      },
...
      "SessionInput": {
        "properties": {
          "credential": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2048
              },
              {
                "type": "null"
              }
            ],
            "title": "Credential"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "SessionInput"
      },
      "SessionView": {
        "properties": {
          "csrf_token": {
            "type": "string",
            "title": "Csrf Token"
          },
          "expires_at": {
            "type": "string",
            "title": "Expires At"
          },
          "mode": {
            "type": "string",
            "enum": [
              "mock",
              "live"
            ],
            "title": "Mode"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "csrf_token",
          "expires_at",
          "mode"
        ],
        "title": "SessionView"
      },
      "Snapshot": {
        "properties": {
          "schema_version": {
            "type": "string",
            "const": "2",
            "title": "Schema Version",
            "default": "2"
          },
          "agreement_id": {
            "type": "string",
            "title": "Agreement Id"
          },
          "product_id": {
            "type": "string",
            "title": "Product Id"
          },
          "seller_id": {
            "type": "string",
            "title": "Seller Id"
          },
          "product_version": {
            "type": "integer",
            "title": "Product Version"
          },
          "seller_version": {
            "type": "integer",
            "title": "Seller Version"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "ram_gb": {
            "type": "integer",
            "title": "Ram Gb"
          },
          "ssd_gb": {
            "type": "integer",
            "title": "Ssd Gb"
          },
          "item_price_krw": {
            "type": "integer",
            "maximum": 100000000.0,
            "exclusiveMinimum": 0.0,
            "title": "Item Price Krw"
          },
          "shipping_fee_krw": {
            "type": "integer",
            "maximum": 100000000.0,
            "minimum": 0.0,
            "title": "Shipping Fee Krw"
          },
          "fee_krw": {
            "type": "integer",
            "maximum": 100000000.0,
            "minimum": 0.0,
            "title": "Fee Krw"
          },
          "total_krw": {
            "type": "integer",
            "maximum": 100000000.0,
            "exclusiveMinimum": 0.0,
            "title": "Total Krw"
          },
          "delivery_date": {
            "type": "string",
            "format": "date",
            "title": "Delivery Date"
          },
          "policy_version": {
            "type": "integer",
            "title": "Policy Version"
          },
          "nonce": {
            "type": "string",
            "title": "Nonce"
          },
          "expiry": {
            "type": "string",
            "format": "date-time",
            "title": "Expiry"
          },
          "source": {
            "type": "string",
            "const": "simulated",
            "title": "Source"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "agreement_id",
          "product_id",
          "seller_id",
          "product_version",
          "seller_version",
          "name",
          "ram_gb",
          "ssd_gb",
          "item_price_krw",
          "shipping_fee_krw",
          "fee_krw",
          "total_krw",
          "delivery_date",
          "policy_version",
          "nonce",
          "expiry",
          "source"
        ],
        "title": "Snapshot"
      },
      "State": {
        "type": "string",
        "enum": [
          "DRAFT",
          "POLICY_CONFIRMED",
          "NEGOTIATING",
          "AWAITING_APPROVAL",
          "BLOCKED",
          "REJECTED",
          "RECORDING",
          "RECORDED",
          "MOCK_RECORDED",
          "CHAIN_FAILED"
        ],
        "title": "State"
      },
...
    },
    "securitySchemes": {
      "APIKeyCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "dealbattle_session"
      }
    }
  }
}